# docs/swagger.yaml
definitions:
...
  model.MySQLTime:
    properties:
      time.Time:
        type: string
    type: object
  model.Review:
    properties:
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      description:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      rating:
        type: integer
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
      user_id:
        type: integer
    type: object
  model.Role:
    properties:
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      id:
        type: integer
      role_name:
        type: string
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
    type: object
...
info:
  contact: {}
paths:
...
  /review:
    get:
      description: Get a list of all reviews
      responses:
        "200":
          description: List of reviews
          schema:
            items:
              $ref: '#/definitions/model.Review'
            type: array
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all reviews
    post:
      description: Update an existing review for a game by the authenticated user
      parameters:
      - description: Review object that needs to be updated
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/model.Review'
      responses:
        "200":
          description: Review updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Rating should be 0-10" (when the updated review rating is greater
            than 10)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found" (when the specified review ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update review
  /review/{id}:
    delete:
      description: Delete a review by its ID
      parameters:
      - description: Review ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Review successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid review ID" (when the review ID in the URL path is not
            a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Review not found" (when the specified review ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete review by ID
  /role:
    get:
      description: Get a list of all roles
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/model.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all roles
    post:
      description: Create a new role with the specified role name
      parameters:
      - description: Role object that needs to be created
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      responses:
        "200":
          description: Role created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Role already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create new role
  /role/{id}:
    delete:
      description: Delete a role by its ID
      parameters:
      - description: Role ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Role successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid role ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete role by ID
...
swagger: "2.0"